input: fingerprint: rename goodix_fod driver to goodix_fp
The goal is to make a common driver for goofix fingerprint driver
that replaces goodix_fod and goodix_ta.
