+++ arch/arm64/configs/grus_defconfig
@@ -319,7 +319,7 @@ CONFIG_INPUT_HBTP_INPUT=y
 CONFIG_INPUT_QPNP_POWER_ON=y
 CONFIG_INPUT_UINPUT=y
 CONFIG_INPUT_FINGERPRINT=y
-CONFIG_FINGERPRINT_GOODIX_FOD=y
+CONFIG_FINGERPRINT_GOODIX_FP=y
 # CONFIG_SERIO_SERPORT is not set
 # CONFIG_VT is not set
 # CONFIG_LEGACY_PTYS is not set
