+++ drivers/input/fingerprint/Kconfig
@@ -13,6 +13,6 @@ if INPUT_FINGERPRINT
 
 source "drivers/input/fingerprint/fpc_tee/Kconfig"
 source "drivers/input/fingerprint/goodix_ta/Kconfig"
-source "drivers/input/fingerprint/goodix_fod/Kconfig"
+source "drivers/input/fingerprint/goodix_fp/Kconfig"
 
 endif
