+++ drivers/input/fingerprint/Makefile
@@ -6,4 +6,4 @@
 
 obj-$(CONFIG_FINGERPRINT_FPC_TEE)	+= fpc_tee/
 obj-$(CONFIG_FINGERPRINT_GOODIX_TA)	+= goodix_ta/
-obj-$(CONFIG_FINGERPRINT_GOODIX_FOD)	+= goodix_fod/
+obj-$(CONFIG_FINGERPRINT_GOODIX_FP)	+= goodix_fp/
