--- drivers/input/fingerprint/goodix_fod/Makefile
@@ -1,1 +1,0 @@
-obj-$(CONFIG_FINGERPRINT_GOODIX_FOD) += gf_spi.o platform.o netlink.o
