+++ drivers/input/fingerprint/goodix_fp/Makefile
@@ -1,0 +1,1 @@
+obj-$(CONFIG_FINGERPRINT_GOODIX_FP) += driver.o platform.o netlink.o
