+++ drivers/input/fingerprint/goodix_fp/platform.c
@@ -14,7 +14,7 @@
 #include <linux/timer.h>
 #include <linux/err.h>
 
-#include "gf_spi.h"
+#include "goodix_fp.h"
 
 #if defined(USE_SPI_BUS)
 #include <linux/spi/spi.h>
